Revert "fix(time): propagate widened Long return types to bridge and planner adapter"
This reverts commit 40061812859083a6ec730ea1a1b953ae9daf0d8f.

+++ sources/Android/android-communications/library/src/sdk/java/com/polar/sdk/impl/utils/PolarRuntimePlannerAdapter.kt
@@ -1912,7 +1912,7 @@ internal object PolarRuntimePlannerAdapter {
         ).toPlanned()
     }
 
-    fun millisToNanos(milliseconds: Int): Long {
+    fun millisToNanos(milliseconds: Int): Int {
         return PolarTimeUtils.millisToNanos(milliseconds)
     }
 
@@ -1920,7 +1920,7 @@ internal object PolarRuntimePlannerAdapter {
         return PolarTimeUtils.secondsToMinutes(seconds)
     }
 
-    fun durationMillis(hours: Int, minutes: Int, seconds: Int, millis: Int): Long {
+    fun durationMillis(hours: Int, minutes: Int, seconds: Int, millis: Int): Int {
         return PolarTimeUtils.durationToMillis(PolarDurationFields(hours, minutes, seconds, millis))
     }
